==> notebooks/1_business_understanding.ipynb <==
## Individual Project - Avocado Dataset
### Author: Aryan Khanjani (563751@student.saxion.nl)
### Course: DataOps Saxion University of Applied Sciences 2025
</br>

## Introduction

As a small business owner of a avocado online shop, I am affected by the constant price fluctuations in the market. In order to mitiagte the risk of buying my merchandise at a higher price or underselling them, I decided to build a simple model that can predict the average price of avocados for the next month. This can help me gain insight into when it is best to buy or sell avocados. This way, I can increase my revenue and survive in the economy.

> The actual purpose of this project is to get familiar with and gain hands-on experience with data science using an Avocado Price dataset.

**Source**: [Kaggle](https://www.kaggle.com/datasets/neuromusic/avocado-prices)

Throughout this project, I will be using the CRISP-DM methodology to form a well-structured project:

1. Business Understanding
2. Data Understanding
3. Data Preparation
4. Modeling
5. Business Evaluation

---

## 1 Business Understanding

### 1.1 Business Goals

The business goals of this project is to predict the price of Avocados for the next month using historical data with a accuracy of at least 70%.

### 1.2 Risks & Contingencies

- The provided dataset is not up to date with its description. The dataset is said to have been expanded since 2013, however the earliest record in the dataset is from 2015.
- 5 week time constraint does not allow to train a model that takes long hours to train or to fine-tune the model.
- Computer's performance is not powerful enough to handle large datasets or to train a model that takes long hours to train.
- Model is missing Exogenous features which can cause potential accuracy issues.
- I am the only one working on this project, so I will not be able to increase the scope of the project, or possibly finish the current scope.

### 1.3 Constraints:

- As the dataset is not very large, removing any data (e.g. outliers) should be done as little as possible.
- Will not include any external exogenous variables in the model due to time constraints (e.g. weather data, holidays, etc.).
- Will not perform refitting, or versioning for the model as these are not the primary focus in this project.
- Will not tune the model or perform gird search due to time constraints.

### 1.4 Terminology:

- PLU (Price Look-Up code): A standardized code used in supermarkets to identify a product.

### 1.5 Data Mining Success Criteria

- Predict the average avocado price for the next month, measured by Mean Absolute Error (MAE) to be less than 0.2.
- The model should be computationally efficient, running in under 1 minutes on a standard laptop.
- The model must use at most 70% of the data as training dataset, leaving 30% for testing.

### 1.6 Research Questions

- How to handle outliers?
- How to handle missing values?
- How to handle categorical variables?
- Is it necessary to scale the data and if so, how to do it?
- How to evaluate the model?
- How to select the best model?
- Are there any other ways to process the data further than what I have learnt in class?
