==> notebooks/5_evaluation.ipynb <==
## 5 Business Evaluation

---

### Model Performance vs. Business Goals

**SARIMAX Model**

- **Target MAE:** < 0.2
- **Achieved MAE:** 0.1395
- **Conclusion:** Model meets accuracy requirements.

**Prophet Model**

- **Target MAE:** < 0.2
- **Achieved MAE:** 0.3901
- **Conclusion:** Model <em>does not</em> meet the accuracy requirements.

The trained SARIMAX model is able to predict the price of avocados for the next month with an MAE of 0.1395, using all the prepared features with an accuracy of 86.05%.

### Interpretation of SARIMAX Model Results

- Prices show seasonal trends.
- Forecasts almost closely align with historical data, making them relatively reliable.

### Business Impact Analysis

- Business can now optimize inventory purchases with a 1-month lead time.

### Risks & Limitations

- **Risk:** Model does not include exogenous variables that are not available in the dataset such as inflation rates, holidays, or transportation costs.
- **Limitation:** Only 3 years of data is used for training the model, which may not be enough to capture long-term trends.
- **Mitigation:** Add external economic indicators to the data and upon signs of correlation, include them in the model.
